ref: Implement Event.add_exception() to replace capture_error()
