+++ src/sentry/disabled_event.h
@@ -49,6 +49,8 @@ public:
 	virtual void remove_tag(const String &p_key) override {}
 	virtual String get_tag(const String &p_key) override { return ""; }
 
+	virtual void add_exception(const String &p_type, const String &p_value, const Vector<StackFrame> &p_frames) override {}
+
 	virtual bool is_crash() const override { return false; }
 };
